v5.x-refactor-tests: removed CostingRules.scala
